pytorch-cifar-master/requirements.txt
numpy==1.26.4
torchvision==0.20.1
timm==1.0.13
pillow==11.1.0
yaml==0.2.5
pyyaml==6.0.2
matplotlib==3.10.0
tqdm==4.67.1